fix(webapp): harden docker entrypoint cleanup and API_URL placeholder

+++ front/docker-entrypoint.sh
@@ -6,13 +6,13 @@ HTML_DIR=/usr/share/angie/html
 SRC_DIR=/usr/local/src/ferriskey
 CONFIG_FILE="$HTML_DIR/config.json"
 
-rm -rf "$HTML_DIR"/*
+rm -rf -- "${HTML_DIR:?}/"* "${HTML_DIR:?}/".[!.]* "${HTML_DIR:?}/"..?*
 cp -r "$SRC_DIR"/* "$HTML_DIR"
 
 if [ -f "$CONFIG_FILE" ]; then
   api_url="${API_URL:-}"
   escaped_api_url=$(printf '%s' "$api_url" | sed -e 's/[\/&]/\\&/g')
-  placeholder='${API_URL}'
+  placeholder="\${API_URL}"
   sed -i "s|$placeholder|$escaped_api_url|g" "$CONFIG_FILE"
 fi
